ESP32 IP Address
this is part of the code used to create the wifi server for the esp32. if you replace the ssid and password with an actual wifi serer (ex a hotspot), then the esp32 will connect to that wifi and give you an IP address. 

since we're looking to send data to the esp32, it seems we need to ensure the esp32 creates a wifi for us to connect to and then we can use the IP address and add our ncessary inputs from there...