Create WiFI from yt vid ESP32
this is code we got from a YouTube video guy where he walked through the process of connecting to a wifi server and sending data to the server.  in the video, it worked for the creator but the code for us wasn't working. it couldn't even connect to a wifi address :(